Refactor namespaces and update entity properties
- Removed `StileStream.Wms.SharedKernel.Infrastructure` project from `StileStreamWms.sln`.
- Removed `CreatedAt`, `UpdatedAt`, `CreatedBy`, and `UpdatedBy` properties from `Product` class in `Product.cs`.
- Updated `InventoryServiceDbContext` to include `OutboxMessages.Entities` and `Products` namespaces.
- Moved `OutboxMessageConfiguration` to `OutboxMessages.Configurations` namespace.
- Moved `OutboxMessage` to `OutboxMessages.Entities` namespace.
- Moved `ProductsConfiguration` to `Products.Configurations` namespace.
- Moved `ProductEntityExtension` to `Products.Extensions` namespace.
- Moved `ProductEntity` to `Products` namespace and made it inherit from `EntityBase`; removed `CreatedAt`, `CreatedBy`, `UpdatedAt`, and `UpdatedBy` properties.
- Moved `ProductRepository` to `Products.Repositories` namespace.
- Updated `DatabaseServiceCollectionExtension` to include `Products.Repositories` namespace.
- Updated `StileStream.Wms.Inventory.Infrastructure.csproj` to remove `Data\NewFolder1\` and `Data\NewFolder\` folders and add a project reference to `StileStream.Wms.SharedKernel.Infrastructure.Data`.
- Added `CreatedAt`, `UpdatedAt`, `CreatedBy`, and `UpdatedBy` properties to `Entity` class in `Entity.cs`.
- Reformatted `IEntity` interface in `IEntity.cs`.
- Removed `StileStream.Wms.SharedKernel.Infrastructure.csproj`.
- Added new `EntityBase` class in `EntityBase.cs` with `CreatedAt`, `CreatedBy`, `UpdatedAt`, and `UpdatedBy` properties.
