+++ StileStreamWms/src/Inventory/StileStream.Wms.Inventory.Infrastructure/Data/Products/EntityBase.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace StileStream.Wms.SharedKernel.Infrastructure.Data.EntityBase;
+
+public class EntityBase
+{
+    public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
+    public string CreatedBy { get; set; } = "system";
+    public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
+    public string UpdatedBy { get; set; } = "system";
+}
