+++ StileStreamWms/src/Inventory/StileStream.Wms.Inventory.Infrastructure/StileStream.Wms.Inventory.Infrastructure.csproj
@@ -7,6 +7,15 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="Data\NewFolder1\**" />
+    <Compile Remove="Data\NewFolder\**" />
+    <EmbeddedResource Remove="Data\NewFolder1\**" />
+    <EmbeddedResource Remove="Data\NewFolder\**" />
+    <None Remove="Data\NewFolder1\**" />
+    <None Remove="Data\NewFolder\**" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.8" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="8.0.8">
       <PrivateAssets>all</PrivateAssets>
@@ -21,12 +30,8 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\SharedKernel\StileStream.Wms.SharedKernel.Domain\StileStream.Wms.SharedKernel.Domain.csproj" />
+    <ProjectReference Include="..\..\SharedKernel\StileStream.Wms.SharedKernel.Infrastructure.Data\StileStream.Wms.SharedKernel.Infrastructure.Data\StileStream.Wms.SharedKernel.Infrastructure.Data.csproj" />
     <ProjectReference Include="..\StileStream.Wms.Inventory.Domain\StileStream.Wms.Inventory.Domain.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Data\NewFolder1\" />
-    <Folder Include="Data\NewFolder\" />
   </ItemGroup>
 
 </Project>
