+++ StileStreamWms/src/SharedKernel/StileStream.Wms.SharedKernel.Domain/Events/Entity.cs
@@ -5,6 +5,10 @@ namespace StileStream.Wms.SharedKernel.Domain.Events;
 public abstract class Entity : IEntity
 {
     private readonly List<IDomainEvent> _domainEvents = [];
+    public DateTime CreatedAt { get; protected set; }
+    public DateTime UpdatedAt { get; protected set; }
+    public string CreatedBy { get; protected set; } = string.Empty;
+    public string UpdatedBy { get; protected set; } = string.Empty;
 
     public IReadOnlyList<IDomainEvent> DomainEvents => _domainEvents;
