--- StileStreamWms/src/SharedKernel/StileStream.Wms.SharedKernel.Infrastructure/StileStream.Wms.SharedKernel.Infrastructure.csproj
@@ -1,17 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.8" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\StileStream.Wms.SharedKernel.Domain\StileStream.Wms.SharedKernel.Domain.csproj" />
-  </ItemGroup>
-
-</Project>
